+++ gui/history.css
