+++ gui/settings.js
