+++ src/emulator/human.py
