+++ src/search/__init__.py
@@ -1,0 +1,4 @@
+from .engine import SearchEngine
+from .history import HistoryManager
+
+__all__ = ["SearchEngine", "HistoryManager"]
